Build Time Improvement (#1067)
* 🚧 build(VSecM): parallelize building images


* minor

* minor

* move vendor

* removed proto generation from bundle.sh

* even faster

* typo

* code cleanup

---------
