+++ core/constants/audit/event.go
@@ -12,16 +12,18 @@ package audit
 
 type Event string
 
-const Enter Event = "vsecm-enter"
+const BadPayload Event = "vsecm-bad-payload"
+const BadPeerSvid Event = "vsecm-bad-peer-spiffeid"
 const BadSpiffeId Event = "vsecm-bad-spiffeid"
 const BrokenBody Event = "vsecm-broken-body"
-const RequestTypeMismatch Event = "vsecm-request-type-mismatch"
-const BadPeerSvid Event = "vsecm-bad-peer-spiffeid"
-const NoSecret Event = "vsecm-no-secret"
-const Ok Event = "vsecm-ok"
-const NoWorkloadId Event = "vsecm-no-wl-id"
-const NoValue Event = "vsecm-no-value"
-const EncryptionFailed Event = "vsecm-encryption-failed"
 const DecryptionFailed Event = "vsecm-decryption-failed"
-const BadPayload Event = "vsecm-bad-payload"
+const EncryptionFailed Event = "vsecm-encryption-failed"
+const Enter Event = "vsecm-enter"
+const NoSecret Event = "vsecm-no-secret"
+const NoValue Event = "vsecm-no-value"
+const NotWorkload = "vsecm-not-workload"
+const NoWorkloadId Event = "vsecm-no-wl-id"
+const NotSentinel = "vsecm-not-sentinel"
+const Ok Event = "vsecm-ok"
+const RequestTypeMismatch Event = "vsecm-request-type-mismatch"
 const RootKeyNotSet = "vsecm-root-key-not-set"
