+++ core/constants/val/val.go
@@ -18,6 +18,7 @@ const TimeNever = "never"
 const BlankRootKey = "{}"
 
 const Ok = "OK"
+const NotOk = "NOK!"
 
 // JsonEmpty is a constant string representing an empty value.
 // This value is generated by the go templating engine
