+++ docs/templates/404.html
@@ -1,0 +1,22 @@
+{#
+# /*
+# |    Protect your secrets, protect your sensitive data.
+# :    Explore VMware Secrets Manager docs at https://vsecm.com/
+# </
+# <>/  keep your secrets... secret
+# >/
+# <>/' Copyright 2023-present VMware Secrets Manager contributors.
+# >/'  SPDX-License-Identifier: BSD-2-Clause
+# */
+#}
+{% extends "index.html" %}
+{% block content %}
+<div class="not-found" style="background:#2e3440;margin:1em auto;padding:1em;border:8px;text-align:center;max-width:620px;">
+    <p><span style="font-weight:bold;font-size:120px">404</span><br>
+        We couldn't find what you were looking for.</p>
+    <p>Please <a href="/" style="color:#ffffff;text-decoration:underline;">Check out <strong>VSecM Home Page</strong></a>
+      for further information, documentation, and use case examples
+      about <strong>VMware Secrets Manager</strong>.
+    </p>
+</div>
+{% endblock %}
