+++ docs/templates/anchor-link.html
@@ -1,3 +1,14 @@
+{#
+# /*
+# |    Protect your secrets, protect your sensitive data.
+# :    Explore VMware Secrets Manager docs at https://vsecm.com/
+# </
+# <>/  keep your secrets... secret
+# >/
+# <>/' Copyright 2023-present VMware Secrets Manager contributors.
+# >/'  SPDX-License-Identifier: BSD-2-Clause
+# */
+#}
 <a class="zola-anchor" href="#{{ id }}" aria-label="Anchor link for: {{ id }}" style="position:relative; top: -4px;">
   <img src="{{ get_url(path="vsecm-32.png") }}" alt="Link" style="vertical-align:middle">
 </a>
