+++ docs/templates/documentation.html
@@ -1,3 +1,14 @@
+{#
+# /*
+# |    Protect your secrets, protect your sensitive data.
+# :    Explore VMware Secrets Manager docs at https://vsecm.com/
+# </
+# <>/  keep your secrets... secret
+# >/
+# <>/' Copyright 2023-present VMware Secrets Manager contributors.
+# >/'  SPDX-License-Identifier: BSD-2-Clause
+# */
+#}
 {% extends "index.html" %}
 
 {% block extra_nav_class %}{% endblock extra_nav_class %}
