+++ docs/templates/page.html
@@ -1,3 +1,14 @@
+{#
+# /*
+# |    Protect your secrets, protect your sensitive data.
+# :    Explore VMware Secrets Manager docs at https://vsecm.com/
+# </
+# <>/  keep your secrets... secret
+# >/
+# <>/' Copyright 2023-present VMware Secrets Manager contributors.
+# >/'  SPDX-License-Identifier: BSD-2-Clause
+# */
+#}
 {% extends "documentation.html" %}
 
 {% block title %}{{ page.title }} | {{ super() }} {% endblock title %}
