+++ makefiles/VSecMBuild.mk
@@ -9,42 +9,65 @@
 # */
 
 serve-docs:
-	./hack/serve-docs.sh
+	./hack/web-serve.sh
+
+sync-docs:
+	./hack/web-sync.sh
+
+bundle-all-local: \
+	inspector-bundle \
+	keygen-bundle \
+	example-sdk-bundle \
+	example-sidecar-bundle \
+	example-multiple-secrets-bundle \
+	example-init-container-bundle \
+	keystone-bundle-ist \
+	safe-bundle-ist \
+	sidecar-bundle-ist \
+	sentinel-bundle-ist \
+	init-container-bundle-ist
+
+bundle-all: \
+	inspector-bundle \
+	keygen-bundle \
+	example-sdk-bundle \
+	example-sidecar-bundle \
+	example-multiple-secrets-bundle \
+	example-init-container-bundle \
+	keystone-bundle-ist \
+	keystone-bundle-ist-fips \
+	safe-bundle-ist \
+	safe-bundle-ist-fips \
+	sidecar-bundle-ist \
+	sidecar-bundle-ist-fips \
+	sentinel-bundle-ist \
+	sentinel-bundle-ist-fips \
+	init-container-bundle-ist \
+	init-container-bundle-ist-fips
+
+push-all: \
+	inspector-push \
+	keygen-push \
+	example-sidecar-push \
+	example-sdk-push \
+	example-multiple-secrets-push \
+	example-init-container-push \
+	keystone-push-ist \
+	keystone-push-ist-fips \
+	safe-push-ist \
+	safe-push-ist-fips \
+	sidecar-push-ist \
+	sidecar-push-ist-fips \
+	sentinel-push-ist \
+	sentinel-push-ist-fips \
+	init-container-push-ist \
+	init-container-push-ist-fips
 
 # Builds everything and pushes to public DockerHub registries.
-build: \
-	inspector-bundle \
-	inspector-push \
-	keygen-bundle \
-	keygen-push \
-	example-sidecar-bundle \
-	example-sidecar-push \
-	example-sdk-bundle \
-	example-sdk-push \
-	example-multiple-secrets-bundle \
-	example-multiple-secrets-push \
-	example-init-container-bundle \
-	example-init-container-push \
-	keystone-bundle-ist \
-	keystone-push-ist \
-	keystone-bundle-ist-fips \
-	keystone-push-ist-fips \
-	safe-bundle-ist \
-	safe-push-ist \
-	safe-bundle-ist-fips \
-	safe-push-ist-fips \
-	sidecar-bundle-ist \
-	sidecar-push-ist \
-	sidecar-bundle-ist-fips \
-	sidecar-push-ist-fips \
-	sentinel-bundle-ist \
-	sentinel-push-ist \
-	sentinel-bundle-ist-fips \
-	sentinel-push-ist-fips \
-	init-container-bundle-ist \
-	init-container-push-ist \
-	init-container-bundle-ist-fips \
-	init-container-push-ist-fips
+build:
+	./hack/generate-proto-files.sh
+	@$(MAKE) -j$(CPU) bundle-all
+	@$(MAKE) -j$(CPU) push-all
 
 # Login to the public EKS registry.
 login-eks:
@@ -95,31 +118,25 @@ build-eks: \
 	init-container-bundle-ist-fips \
 	init-container-push-ist-fips-eks
 
+push-all-local: \
+	inspector-push-local \
+	keygen-push-local \
+	example-sidecar-push-local \
+	example-sdk-push-local \
+	example-multiple-secrets-push-local \
+	example-init-container-push-local \
+	keystone-push-ist-local \
+	safe-push-ist-local \
+	sidecar-push-ist-local \
+	sentinel-push-ist-local \
+	init-container-push-ist-local
+
 
 # Builds everything and pushes to the local registry.
-build-local: \
-	inspector-bundle \
-	inspector-push-local \
-	keygen-bundle \
-	keygen-push-local \
-	example-sidecar-bundle \
-	example-sidecar-push-local \
-	example-sdk-bundle \
-	example-sdk-push-local \
-	example-multiple-secrets-bundle \
-	example-multiple-secrets-push-local \
-	example-init-container-bundle \
-	example-init-container-push-local \
-	keystone-bundle-ist \
-	keystone-push-ist-local \
-	safe-bundle-ist \
-	safe-push-ist-local \
-	sidecar-bundle-ist \
-	sidecar-push-ist-local \
-	sentinel-bundle-ist \
-	sentinel-push-ist-local \
-	init-container-bundle-ist \
-	init-container-push-ist-local
+build-local:
+	./hack/generate-proto-files.sh
+	@$(MAKE) -j$(CPU) bundle-all-local
+	@$(MAKE) -j$(CPU) push-all-local
 
 build-essentials-local: \
 	keygen-bundle \
